add admin, user features in project3

+++ project3/TodoApp/routers/todos.py
@@ -5,6 +5,7 @@ from typing import Annotated
 from sqlalchemy.orm import Session
 from starlette import status
 from pydantic import BaseModel, Field
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -17,6 +18,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -25,26 +27,39 @@ class TodoRequest(BaseModel):
     complete: bool
 
 @router.get('/', status_code=status.HTTP_200_OK)
-def read_all(db: db_dependency): # this function relies on db, so Depends called Dependency Injection
-    return db.query(Todos).all()     
+def read_all(user: user_dependency, db: db_dependency): # this function relies on db, so Depends called Dependency Injection
+   
+   if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+   return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()    
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first() # .first() because we know that id is primary key and unique
+def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first() # .first() because we know that id is primary key and unique
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="ID not found")
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.dict())
+def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
+
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+    todo_model = Todos(**todo_request.dict(), owner_id = user.get('id'))
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+def update_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     
     if todo_model is None:
         raise HTTPException(status_code=404, detail="ID not found")
@@ -58,9 +73,12 @@ def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int = Pat
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="ID not found")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
